refactor: separate test methods
