+++ frontend/appflowy_tauri/src-tauri/Cargo.toml
@@ -38,7 +38,7 @@ custom-protocol = ["tauri/custom-protocol"]
 # Run the script:
 # scripts/tool/update_client_api_rev.sh  new_rev_id
 # ⚠️⚠️⚠️️
-client-api = { git = "https://github.com/AppFlowy-IO/AppFlowy-Cloud", rev = "49f994488a1d72905a0695f9bdcd801bfb0c577f" }
+client-api = { git = "https://github.com/AppFlowy-IO/AppFlowy-Cloud", rev = "f5fd2643093c5239e46c5f574bbd152a11c15a78" }
 # Please use the following script to update collab.
 # Working directory: frontend
 #
