+++ frontend/rust-lib/event-integration/tests/folder/local_test/script.rs
@@ -1,7 +1,7 @@
 use collab_folder::core::ViewLayout;
 
 use event_integration::event_builder::EventBuilder;
-use event_integration::FlowyCoreTest;
+use event_integration::EventIntegrationTest;
 use flowy_folder2::entities::icon::{UpdateViewIconPayloadPB, ViewIconPB};
 use flowy_folder2::entities::*;
 use flowy_folder2::event_map::FolderEvent::*;
@@ -64,7 +64,7 @@ pub enum FolderScript {
 }
 
 pub struct FolderTest {
-  pub sdk: FlowyCoreTest,
+  pub sdk: EventIntegrationTest,
   pub all_workspace: Vec<WorkspacePB>,
   pub workspace: WorkspacePB,
   pub parent_view: ViewPB,
@@ -75,7 +75,7 @@ pub struct FolderTest {
 
 impl FolderTest {
   pub async fn new() -> Self {
-    let sdk = FlowyCoreTest::new();
+    let sdk = EventIntegrationTest::new();
     let _ = sdk.init_user().await;
     let workspace = create_workspace(&sdk, "FolderWorkspace", "Folder test workspace").await;
     let parent_view = create_app(&sdk, &workspace.id, "Folder App", "Folder test app").await;
@@ -201,7 +201,7 @@ impl FolderTest {
     }
   }
 }
-pub async fn create_workspace(sdk: &FlowyCoreTest, name: &str, desc: &str) -> WorkspacePB {
+pub async fn create_workspace(sdk: &EventIntegrationTest, name: &str, desc: &str) -> WorkspacePB {
   let request = CreateWorkspacePayloadPB {
     name: name.to_owned(),
     desc: desc.to_owned(),
@@ -215,7 +215,10 @@ pub async fn create_workspace(sdk: &FlowyCoreTest, name: &str, desc: &str) -> Wo
     .parse::<WorkspacePB>()
 }
 
-pub async fn read_workspace(sdk: &FlowyCoreTest, workspace_id: Option<String>) -> Vec<WorkspacePB> {
+pub async fn read_workspace(
+  sdk: &EventIntegrationTest,
+  workspace_id: Option<String>,
+) -> Vec<WorkspacePB> {
   let request = WorkspaceIdPB {
     value: workspace_id,
   };
@@ -241,7 +244,12 @@ pub async fn read_workspace(sdk: &FlowyCoreTest, workspace_id: Option<String>) -
   workspaces
 }
 
-pub async fn create_app(sdk: &FlowyCoreTest, workspace_id: &str, name: &str, desc: &str) -> ViewPB {
+pub async fn create_app(
+  sdk: &EventIntegrationTest,
+  workspace_id: &str,
+  name: &str,
+  desc: &str,
+) -> ViewPB {
   let create_view_request = CreateViewPayloadPB {
     parent_view_id: workspace_id.to_owned(),
     name: name.to_string(),
@@ -263,7 +271,7 @@ pub async fn create_app(sdk: &FlowyCoreTest, workspace_id: &str, name: &str, des
 }
 
 pub async fn create_view(
-  sdk: &FlowyCoreTest,
+  sdk: &EventIntegrationTest,
   app_id: &str,
   name: &str,
   desc: &str,
@@ -288,7 +296,7 @@ pub async fn create_view(
     .parse::<ViewPB>()
 }
 
-pub async fn read_view(sdk: &FlowyCoreTest, view_id: &str) -> ViewPB {
+pub async fn read_view(sdk: &EventIntegrationTest, view_id: &str) -> ViewPB {
   let view_id = ViewIdPB::from(view_id);
   EventBuilder::new(sdk.clone())
     .event(ReadView)
@@ -299,7 +307,7 @@ pub async fn read_view(sdk: &FlowyCoreTest, view_id: &str) -> ViewPB {
 }
 
 pub async fn move_view(
-  sdk: &FlowyCoreTest,
+  sdk: &EventIntegrationTest,
   view_id: String,
   parent_id: String,
   prev_view_id: Option<String>,
@@ -319,7 +327,7 @@ pub async fn move_view(
   assert!(error.is_none());
 }
 pub async fn update_view(
-  sdk: &FlowyCoreTest,
+  sdk: &EventIntegrationTest,
   view_id: &str,
   name: Option<String>,
   desc: Option<String>,
@@ -340,7 +348,7 @@ pub async fn update_view(
     .await;
 }
 
-pub async fn update_view_icon(sdk: &FlowyCoreTest, view_id: &str, icon: Option<ViewIconPB>) {
+pub async fn update_view_icon(sdk: &EventIntegrationTest, view_id: &str, icon: Option<ViewIconPB>) {
   let request = UpdateViewIconPayloadPB {
     view_id: view_id.to_string(),
     icon,
@@ -352,7 +360,7 @@ pub async fn update_view_icon(sdk: &FlowyCoreTest, view_id: &str, icon: Option<V
     .await;
 }
 
-pub async fn delete_view(sdk: &FlowyCoreTest, view_ids: Vec<String>) {
+pub async fn delete_view(sdk: &EventIntegrationTest, view_ids: Vec<String>) {
   let request = RepeatedViewIdPB { items: view_ids };
   EventBuilder::new(sdk.clone())
     .event(DeleteView)
@@ -361,7 +369,7 @@ pub async fn delete_view(sdk: &FlowyCoreTest, view_ids: Vec<String>) {
     .await;
 }
 
-pub async fn read_trash(sdk: &FlowyCoreTest) -> RepeatedTrashPB {
+pub async fn read_trash(sdk: &EventIntegrationTest) -> RepeatedTrashPB {
   EventBuilder::new(sdk.clone())
     .event(ReadTrash)
     .async_send()
@@ -369,7 +377,7 @@ pub async fn read_trash(sdk: &FlowyCoreTest) -> RepeatedTrashPB {
     .parse::<RepeatedTrashPB>()
 }
 
-pub async fn restore_app_from_trash(sdk: &FlowyCoreTest, app_id: &str) {
+pub async fn restore_app_from_trash(sdk: &EventIntegrationTest, app_id: &str) {
   let id = TrashIdPB {
     id: app_id.to_owned(),
   };
@@ -380,7 +388,7 @@ pub async fn restore_app_from_trash(sdk: &FlowyCoreTest, app_id: &str) {
     .await;
 }
 
-pub async fn restore_view_from_trash(sdk: &FlowyCoreTest, view_id: &str) {
+pub async fn restore_view_from_trash(sdk: &EventIntegrationTest, view_id: &str) {
   let id = TrashIdPB {
     id: view_id.to_owned(),
   };
@@ -391,14 +399,14 @@ pub async fn restore_view_from_trash(sdk: &FlowyCoreTest, view_id: &str) {
     .await;
 }
 
-pub async fn delete_all_trash(sdk: &FlowyCoreTest) {
+pub async fn delete_all_trash(sdk: &EventIntegrationTest) {
   EventBuilder::new(sdk.clone())
     .event(DeleteAllTrash)
     .async_send()
     .await;
 }
 
-pub async fn toggle_favorites(sdk: &FlowyCoreTest, view_id: Vec<String>) {
+pub async fn toggle_favorites(sdk: &EventIntegrationTest, view_id: Vec<String>) {
   let request = RepeatedViewIdPB { items: view_id };
   EventBuilder::new(sdk.clone())
     .event(ToggleFavorite)
@@ -407,7 +415,7 @@ pub async fn toggle_favorites(sdk: &FlowyCoreTest, view_id: Vec<String>) {
     .await;
 }
 
-pub async fn read_favorites(sdk: &FlowyCoreTest) -> RepeatedViewPB {
+pub async fn read_favorites(sdk: &EventIntegrationTest) -> RepeatedViewPB {
   EventBuilder::new(sdk.clone())
     .event(ReadFavorites)
     .async_send()
