+++ frontend/rust-lib/event-integration/tests/folder/local_test/subscription_test.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use event_integration::FlowyCoreTest;
+use event_integration::EventIntegrationTest;
 use flowy_folder2::entities::{ChildViewUpdatePB, RepeatedViewPB, UpdateViewPayloadPB};
 use flowy_folder2::notification::FolderNotification;
 
@@ -16,7 +16,7 @@ use crate::util::receive_with_timeout;
 /// 5. Await the notification for workspace view updates with a timeout of 30 seconds.
 /// 6. Ensure that the received views contain the newly created "test_view".
 async fn create_child_view_in_workspace_subscription_test() {
-  let test = FlowyCoreTest::new_with_guest_user().await;
+  let test = EventIntegrationTest::new_with_guest_user().await;
   let workspace = test.get_current_workspace().await.workspace;
   let mut rx = test
     .notification_sender
@@ -40,7 +40,7 @@ async fn create_child_view_in_workspace_subscription_test() {
 
 #[tokio::test]
 async fn create_child_view_in_view_subscription_test() {
-  let test = FlowyCoreTest::new_with_guest_user().await;
+  let test = EventIntegrationTest::new_with_guest_user().await;
   let mut workspace = test.get_current_workspace().await.workspace;
   let workspace_child_view = workspace.views.pop().unwrap();
   let mut rx = test.notification_sender.subscribe::<ChildViewUpdatePB>(
@@ -72,7 +72,7 @@ async fn create_child_view_in_view_subscription_test() {
 
 #[tokio::test]
 async fn delete_view_subscription_test() {
-  let test = FlowyCoreTest::new_with_guest_user().await;
+  let test = EventIntegrationTest::new_with_guest_user().await;
   let workspace = test.get_current_workspace().await.workspace;
   let mut rx = test
     .notification_sender
@@ -94,7 +94,7 @@ async fn delete_view_subscription_test() {
 
 #[tokio::test]
 async fn update_view_subscription_test() {
-  let test = FlowyCoreTest::new_with_guest_user().await;
+  let test = EventIntegrationTest::new_with_guest_user().await;
   let mut workspace = test.get_current_workspace().await.workspace;
   let mut rx = test
     .notification_sender
