+++ frontend/rust-lib/event-integration/tests/user/local_test/user_awareness_test.rs
@@ -1,13 +1,13 @@
 use std::collections::HashMap;
 
 use event_integration::event_builder::EventBuilder;
-use event_integration::FlowyCoreTest;
+use event_integration::EventIntegrationTest;
 use flowy_user::entities::{ReminderPB, RepeatedReminderPB};
 use flowy_user::event_map::UserEvent::*;
 
 #[tokio::test]
 async fn user_update_with_name() {
-  let sdk = FlowyCoreTest::new();
+  let sdk = EventIntegrationTest::new();
   let _ = sdk.sign_up_as_guest().await;
   let mut meta = HashMap::new();
   meta.insert("object_id".to_string(), "".to_string());
